make PaymentSheetPlaygroundActivity an ExternalPaymentMethodConfirmHandler
